dns_inspector_toolkit/requirements.txt

dnspython
python-whois
